Rename cbh_analysis crate to cbh_detect
`cbh_analysis` and the sibling command crate `cbh_analyze` differed by only
two letters and were easy to confuse. Rename the pure detection-engine crate to
`cbh_detect`, which names what it does (change-point / regression detection) and
shares no stem with `analyze`. The command crate keeps the `cbh_analyze` name,
matching its flagship `analyze` subcommand.

Also renames the crate's internal top-level module `analyze` -> `detect`, so
`crate::detect::` no longer reads like a reference to the `cbh_analyze` crate.
Command references in prose (the `analyze` subcommand, `docs/analyze.md`, the
`Analysis*` value types) are left unchanged.

Updates the workspace dependency declaration, the two dependents (`cbh_analyze`,
`cbh_render`), `release-plz.toml`, `Mutants.psm1`, and the docs, keeping the
alphabetical dependency and package ordering intact.
