+++ Cargo.toml
@@ -22,12 +22,12 @@ azure_identity = { version = "1.0.0", default-features = false }
 azure_storage_blob = { version = "1.0.0", default-features = false }
 base64 = { version = "0.22.1", default-features = false, features = ["std"] }
 cargo-freeze-deps = { version = "0.1.3", path = "packages/cargo-freeze-deps", default-features = false }
-cbh_analysis = { version = "=0.0.3", path = "packages/cbh_analysis", default-features = false }
 cbh_analyze = { version = "=0.0.3", path = "packages/cbh_analyze", default-features = false }
 cbh_cli = { version = "=0.0.3", path = "packages/cbh_cli", default-features = false }
 cbh_codec = { version = "=0.0.3", path = "packages/cbh_codec", default-features = false }
 cbh_command = { version = "=0.0.3", path = "packages/cbh_command", default-features = false }
 cbh_config = { version = "=0.0.3", path = "packages/cbh_config", default-features = false }
+cbh_detect = { version = "=0.0.3", path = "packages/cbh_detect", default-features = false }
 cbh_diag = { version = "=0.0.3", path = "packages/cbh_diag", default-features = false }
 cbh_engines = { version = "=0.0.3", path = "packages/cbh_engines", default-features = false }
 cbh_git = { version = "=0.0.3", path = "packages/cbh_git", default-features = false }
