+++ packages/cbh_analyze/src/dataset.rs
@@ -6,8 +6,8 @@ use std::sync::Arc;
 use std::time::Instant;
 
 use anyspawn::Spawner;
-use cbh_analysis::{AnalysisMode, BlessingPlacement, Series, SeriesFilter};
 use cbh_config::Config;
+use cbh_detect::{AnalysisMode, BlessingPlacement, Series, SeriesFilter};
 use cbh_diag::{Reporter, ReporterExt, count_noun};
 use cbh_git::GitHistory;
 use cbh_model::{BenchmarkIdPrefix, BlessingRecord, DiscriminantSet, StorageKey};
