+++ packages/cbh_analyze/src/lib.rs
@@ -38,7 +38,7 @@ mod selection;
 mod window;
 
 pub use bless::{bless, unbless};
-pub(crate) use cbh_analysis::{Series, SeriesFilter, apply_blessings};
+pub(crate) use cbh_detect::{Series, SeriesFilter, apply_blessings};
 pub(crate) use cbh_render::{ReportFormat, format_value};
 pub(crate) use dataset::{empty_history_hint, select_dataset};
 pub use examine::execute as examine;
