+++ packages/cbh_analyze/src/list.rs
@@ -906,7 +906,7 @@ mod tests {
     /// An inline spawner that runs the load tasks on the calling thread, so the
     /// list tests need no Tokio runtime under `block_on` or Miri.
     fn spawner() -> Spawner {
-        cbh_analysis::testing::synchronous_spawner()
+        cbh_detect::testing::synchronous_spawner()
     }
 
     /// A result set with one record carrying two metrics, so its partition
