+++ packages/cbh_detect/src/detect/discriminant.rs
