+++ packages/cbh_detect/src/detect/parallel.rs
