+++ packages/cbh_detect/src/testing.rs
