+++ packages/cbh_stats/src/lib.rs
@@ -14,7 +14,7 @@
 //! Mann–Whitney, Mann–Kendall, Theil–Sen, Benjamini–Hochberg) for the analysis
 //! detectors, split out of `cargo-bench-history` so this deterministic,
 //! I/O-free, Miri-safe math is cheap to mutation-test in isolation. The
-//! `cbh_analysis` detectors compose these primitives.
+//! `cbh_detect` detectors compose these primitives.
 //!
 //! [`cargo-bench-history`]: https://github.com/folo-rs/folo
